+++ firmware/clockfaces/cw-cf-0x07/Clockface.h
@@ -19,6 +19,7 @@
 #include "fonts/TomThumb.h"
 #include "fonts/FreeSansBold9pt7b.h"
 #include "fonts/nocturno.h"
+#include "fonts/ledDisplay.h"
 #include "PNGRender.h"
 #include "CustomSprite.h"
 #include "CWHttpClient.h"